adaugare service-uri CRUD entitati

+++ solution/src/main/java/com/example/demo/repository/StoreRepo.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepo extends JpaRepository<Store, Integer> {
+public interface StoreRepo extends JpaRepository<Store, Long> {
 }
